Add disassembly of ppc32_microcode.
Copied from the dev/microcode branch of dynamips.

Heavy on comments, the aim was to produce an equivalent microcode.
The contents look equivalent according to objdump output.
It contains lots of debug info which I haven't tried to remove.

The next challenge is to try and make a C source microcode.